requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
dotenv==0.9.9
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
pymongo==4.11.3
python-dotenv==1.1.0
Werkzeug==3.1.3
paypalrestsdk==1.13.3
